refactor!: extract plaintext encrypter into separate module
BREAKING CHANGE: imports from `libp2p/plaintext` should be changed to `@libp2p/plaintext`
