+++ packages/connection-encrypter-plaintext/package.json
@@ -1,0 +1,68 @@
+{
+  "name": "@libp2p/plaintext",
+  "version": "0.0.0",
+  "description": "An insecure connection encrypter",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p/tree/master/packages/connection-encrypter-plaintext#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/libp2p/js-libp2p.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p/issues"
+  },
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "types": "./dist/src/index.d.ts",
+      "import": "./dist/src/index.js"
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "project": true,
+      "sourceType": "module"
+    }
+  },
+  "scripts": {
+    "start": "node dist/src/main.js",
+    "build": "aegir build",
+    "test": "aegir test",
+    "clean": "aegir clean",
+    "generate": "protons ./src/pb/index.proto",
+    "lint": "aegir lint",
+    "test:chrome": "aegir test -t browser --cov",
+    "test:chrome-webworker": "aegir test -t webworker",
+    "test:firefox": "aegir test -t browser -- --browser firefox",
+    "test:firefox-webworker": "aegir test -t webworker -- --browser firefox",
+    "test:node": "aegir test -t node --cov",
+    "dep-check": "aegir dep-check"
+  },
+  "dependencies": {
+    "@libp2p/interface": "^0.1.2",
+    "@libp2p/peer-id": "^3.0.6",
+    "@multiformats/multiaddr": "^12.1.5",
+    "it-handshake": "^4.1.3",
+    "it-length-prefixed": "^9.0.3",
+    "it-map": "^3.0.4",
+    "it-stream-types": "^2.0.1",
+    "protons-runtime": "^5.0.0",
+    "uint8arraylist": "^2.4.3"
+  },
+  "devDependencies": {
+    "@libp2p/interface-compliance-tests": "^4.1.5",
+    "@libp2p/logger": "^3.0.2",
+    "@libp2p/peer-id-factory": "^3.0.8",
+    "aegir": "^41.0.2",
+    "protons": "^7.3.0",
+    "sinon": "^17.0.0"
+  }
+}
