+++ packages/connection-encrypter-plaintext/test/compliance.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 
 import suite from '@libp2p/interface-compliance-tests/connection-encryption'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '../src/index.js'
 
 describe('plaintext compliance', () => {
   suite({
