+++ packages/libp2p/test/addresses/addresses.node.ts
@@ -6,7 +6,7 @@ import { type Multiaddr, multiaddr, protocols } from '@multiformats/multiaddr'
 import { expect } from 'aegir/chai'
 import { pEvent } from 'p-event'
 import sinon from 'sinon'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { createNode } from '../fixtures/creators/peer.js'
 import { AddressesOptions } from './utils.js'
 import type { Libp2pNode } from '../../src/libp2p.js'
