+++ packages/libp2p/test/circuit-relay/discovery.node.ts
@@ -6,7 +6,7 @@ import { expect } from 'aegir/chai'
 import { pEvent } from 'p-event'
 import { circuitRelayServer, type CircuitRelayService, circuitRelayTransport } from '../../src/circuit-relay/index.js'
 import { createLibp2p } from '../../src/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { getRelayAddress, hasRelay, MockContentRouting, mockContentRouting } from './utils.js'
 import type { Libp2p } from '@libp2p/interface'
