+++ packages/libp2p/test/circuit-relay/relay.spec.ts
@@ -11,7 +11,7 @@ import { expect } from 'aegir/chai'
 import { pEvent } from 'p-event'
 import { circuitRelayTransport } from '../../src/circuit-relay/index.js'
 import { createLibp2p } from '../../src/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { hasRelay } from './utils.js'
 import type { Libp2p } from '@libp2p/interface'
 import type { Connection } from '@libp2p/interface/connection'
