+++ packages/libp2p/test/connection-manager/direct.node.ts
@@ -30,7 +30,7 @@ import { defaultComponents, type Components } from '../../src/components.js'
 import { DialQueue } from '../../src/connection-manager/dial-queue.js'
 import { DefaultConnectionManager } from '../../src/connection-manager/index.js'
 import { codes as ErrorCodes } from '../../src/errors.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { createLibp2pNode, type Libp2pNode } from '../../src/libp2p.js'
 import { DefaultTransportManager } from '../../src/transport-manager.js'
 import type { PeerId } from '@libp2p/interface/peer-id'
