+++ packages/libp2p/test/content-routing/dht/operation.node.ts
@@ -9,7 +9,7 @@ import { expect } from 'aegir/chai'
 import pWaitFor from 'p-wait-for'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import { createLibp2p } from '../../../src/index.js'
-import { plaintext } from '../../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { createPeerId } from '../../fixtures/creators/peer.js'
 import { subsystemMulticodecs } from './utils.js'
 import type { Libp2p } from '@libp2p/interface'
