+++ packages/libp2p/test/core/consume-peer-record.spec.ts
@@ -2,7 +2,7 @@
 
 import { webSockets } from '@libp2p/websockets'
 import { multiaddr } from '@multiformats/multiaddr'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { createLibp2pNode, type Libp2pNode } from '../../src/libp2p.js'
 import { createPeerId } from '../fixtures/creators/peer.js'
