+++ packages/libp2p/test/core/events.spec.ts
@@ -4,7 +4,7 @@ import { webSockets } from '@libp2p/websockets'
 import { expect } from 'aegir/chai'
 import { pEvent } from 'p-event'
 import { createLibp2p } from '../../src/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import type { Libp2p } from '@libp2p/interface'
 
 describe('events', () => {
