+++ packages/libp2p/test/core/listening.node.ts
@@ -2,7 +2,7 @@
 
 import { tcp } from '@libp2p/tcp'
 import { expect } from 'aegir/chai'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { createLibp2pNode, type Libp2pNode } from '../../src/libp2p.js'
 import { createPeerId } from '../fixtures/creators/peer.js'
 import type { PeerId } from '@libp2p/interface/peer-id'
