+++ packages/libp2p/test/core/peer-id.spec.ts
@@ -3,7 +3,7 @@
 import { webSockets } from '@libp2p/websockets'
 import { expect } from 'aegir/chai'
 import { createLibp2p, type Libp2p } from '../../src/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 
 describe('peer-id', () => {
   let libp2p: Libp2p
