+++ packages/libp2p/test/fetch/fetch.node.ts
@@ -7,7 +7,7 @@ import { expect } from 'aegir/chai'
 import { codes } from '../../src/errors.js'
 import { type FetchService, fetchService } from '../../src/fetch/index.js'
 import { createLibp2p } from '../../src/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { createPeerId } from '../fixtures/creators/peer.js'
 import type { Libp2p } from '@libp2p/interface'
 import type { PeerId } from '@libp2p/interface/peer-id'
