+++ packages/libp2p/test/fixtures/base-options.ts
@@ -5,7 +5,7 @@ import { webSockets } from '@libp2p/websockets'
 import * as filters from '@libp2p/websockets/filters'
 import mergeOptions from 'merge-options'
 import { circuitRelayTransport } from '../../src/circuit-relay/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import type { Libp2pOptions } from '../../src'
 import type { ServiceMap } from '@libp2p/interface'
