+++ packages/libp2p/test/transports/transport-manager.spec.ts
@@ -13,7 +13,7 @@ import sinon from 'sinon'
 import { DefaultAddressManager } from '../../src/address-manager/index.js'
 import { codes as ErrorCodes } from '../../src/errors.js'
 import { createLibp2p } from '../../src/index.js'
-import { plaintext } from '../../src/insecure/index.js'
+import { plaintext } from '@libp2p/plaintext'
 import { DefaultTransportManager } from '../../src/transport-manager.js'
 import type { Components } from '../../src/components.js'
 import type { Libp2p } from '@libp2p/interface'
